change data generation to multi threads
